backend/requirements.txt
seaborn==0.13.2
pandas==2.2.3
numpy==1.23.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.39
scikit-learn==1.0
scipy==1.10.1
openpyxl==3.1.5
python-dotenv==0.19.1
beautifulsoup4==4.13.3
requests==2.32.3
matplotlib==3.9.4
playwright==1.50.0
pyee==12.1.1
psycopg2==2.9.10